ENH: Make the remaining integer conversions explicit
The findings that are not allocation, copy or I/O lengths: file offsets,
extension bookkeeping, and the ANALYZE orientation byte.

znztell() returns a signed znz_off_t stored in size_t locals, so the
znzseek() arguments are cast to znz_off_t and the arithmetic happens at
the width the function takes. fslio's nvox is widened per operand so the
product of seven int dimensions is computed at 64 bits. XML_Parse() takes
an int length while blen is unsigned; the conversion is now at the call.

The ANALYZE orientation byte is read through unsigned char * rather than
a signed lvalue. That is value-preserving either way -- signed to
unsigned char is defined modulo 256 -- so it states the intent only.

+++ cifti/afni_xml_io.c
@@ -119,7 +119,7 @@ int axio_num_tokens(const char * str, int64_t maxlen)
    if( ! str || ! * str ) return 0;
 
    if( maxlen > 0 ) len = maxlen;
-   else             len = strlen(str);
+   else             len = (int64_t)strlen(str);
 
    ntok = 0;
    intok = 0;
