+++ znzlib/znzlib.c
@@ -148,7 +148,7 @@ size_t znzread(void* buf, size_t size, size_t nmemb, znzFile file)
        /* 0, not gzread's -1: this returns size_t, where -1 is SIZE_MAX. */
        if( nread < 0 ) return 0;
 
-       remain -= nread;
+       remain -= (size_t)nread;
        cbuf += nread;
 
        /* require reading n2read bytes, so we don't get stuck */
@@ -182,7 +182,7 @@ size_t znzwrite(const void* buf, size_t size, size_t nmemb, znzFile file)
        /* gzwrite returns 0 on error, but in case that ever changes... */
        if( nwritten < 0 ) return 0;
 
-       remain -= nwritten;
+       remain -= (size_t)nwritten;
        cbuf += nwritten;
 
        /* require writing n2write bytes, so we don't get stuck */
